Added data loaders to load an individual session or speaker. Added a speakerType to configure which fields to load with data.

+++ GraphQL/Models/Speaker.cs
@@ -4,7 +4,7 @@ namespace GraphQL.Models;
 
 public sealed class Speaker
 {
-    public long Id { get; init; }
+    public int Id { get; init; }
 
     [Required]
     [StringLength(200)]
